Twitter-Analysis/Twitter-Analysis/requirements.txt
Flask==0.12.2
cf-deployment-tracker==1.0.4
cloudant==2.4.0
python-twitter==3.3
watson-developer-cloud==0.26.0